requirements.txt
pillow
numpy
matplotlib
tensorflow
tensorflow_hub
FPDF
simplejson